fix(database): change bigint to integer for consistency with serial

+++ src/main/resources/db/migration/V1_2_2__Client.sql
@@ -1,5 +1,5 @@
 CREATE TABLE public.client (
-    id                      BIGINT          PRIMARY KEY REFERENCES public.user(id) ON DELETE CASCADE,
+    id                      INTEGER         PRIMARY KEY REFERENCES public.user(id) ON DELETE CASCADE,
     phone_number            VARCHAR(255)    NOT NULL UNIQUE,
     passport_number         VARCHAR(255)    NOT NULL UNIQUE,
     identification_number   VARCHAR(255)    NOT NULL UNIQUE
@@ -11,8 +11,8 @@ CREATE TABLE public.client_role (
 );
 
 CREATE TABLE public.client_role_client (
-    role_id         BIGINT          NOT NULL,
-    client_id       BIGINT          NOT NULL,
+    role_id         INTEGER          NOT NULL,
+    client_id       INTEGER         NOT NULL,
 
     PRIMARY KEY (role_id,  client_id),
     FOREIGN KEY (role_id)           REFERENCES public.client_role(id),
