+++ src/main/resources/db/migration/V1_3_3__Enterprise_Account.sql
@@ -1,7 +1,7 @@
 CREATE TABLE public.enterprise_account (
-    id              BIGINT      PRIMARY KEY REFERENCES account(id) ON DELETE CASCADE,
-    enterprise_id   BIGINT      NOT NULL,
-    specialist_id   BIGINT      NOT NULL,
+    id              INTEGER     PRIMARY KEY REFERENCES account(id) ON DELETE CASCADE,
+    enterprise_id   INTEGER     NOT NULL,
+    specialist_id   INTEGER     NOT NULL,
 
     CONSTRAINT fk_enterprise
         FOREIGN KEY (enterprise_id) REFERENCES public.enterprise(id),
