+++ src/main/resources/db/migration/V1_3_5__Salary_Account.sql
@@ -1,8 +1,8 @@
 CREATE TABLE public.salary_account (
-    id                  BIGINT          PRIMARY KEY REFERENCES public.account(id) ON DELETE CASCADE,
-    holder_id           BIGINT          NOT NULL,
+    id                  INTEGER         PRIMARY KEY REFERENCES public.account(id) ON DELETE CASCADE,
+    holder_id           INTEGER         NOT NULL,
     salary              NUMERIC(19,4)   NOT NULL,
-    salary_project_id   BIGINT          NOT NULL,
+    salary_project_id   INTEGER         NOT NULL,
 
     CONSTRAINT fk_holder
         FOREIGN KEY (holder_id) REFERENCES public.client(id),
